Rename config to module to conf.
